+++ src/Pim/Bundle/CatalogBundle/PimCatalogBundle.php
@@ -46,8 +46,17 @@ class PimCatalogBundle extends Bundle
     public function __construct()
     {
         if (class_exists('\Doctrine\ODM\MongoDB\Types\Type')) {
-            \Doctrine\ODM\MongoDB\Types\Type::registerType(self::ODM_ENTITIES_TYPE, 'Pim\Bundle\CatalogBundle\MongoDB\Type\Entities');
-            \Doctrine\ODM\MongoDB\Types\Type::registerType(self::ODM_ENTITY_TYPE, 'Pim\Bundle\CatalogBundle\MongoDB\Type\Entity');
+
+            \Doctrine\ODM\MongoDB\Types\Type::registerType(
+                self::ODM_ENTITIES_TYPE,
+                'Pim\Bundle\CatalogBundle\MongoDB\Type\Entities'
+            );
+
+            \Doctrine\ODM\MongoDB\Types\Type::registerType(
+                self::ODM_ENTITY_TYPE,
+                'Pim\Bundle\CatalogBundle\MongoDB\Type\Entity'
+            );
+
         }
     }