+++ src/Pim/Bundle/DataGridBundle/Datasource/ResultRecord/MongoDbOdm/Hydrator.php
@@ -20,6 +20,7 @@ class Hydrator implements HydratorInterface
     public function hydrate($queryBuilder, $options)
     {
         $localeCode = $options['locale_code'];
+        $scopeCode  = $options['scope_code'];
         $config     = $options['attributes_configuration'];
 
         $query = $queryBuilder->hydrate(false)->getQuery();
@@ -35,12 +36,18 @@ class Hydrator implements HydratorInterface
             $result['id']= $result['_id']->__toString();
             unset($result['_id']);
             $result['dataLocale']= $localeCode;
-            foreach ($result['values'] as $value) {
-                $attribute = $attributes[$value['attribute']];
-                $value['attribute']= $attribute;
-                $result[$attribute['code']]= $value;
+            if (isset($result['values'])) {
+                foreach ($result['values'] as $value) {
+                    $filterValueLocale = isset($value['locale']) && ($value['locale'] !== $localeCode);
+                    $filterValueScope = isset($value['scope']) && ($value['scope'] !== $scopeCode);
+                    if (!$filterValueLocale && !$filterValueScope) {
+                        $attribute = $attributes[$value['attribute']];
+                        $value['attribute']= $attribute;
+                        $result[$attribute['code']]= $value;
+                    }
+                }
+                unset($result['values']);
             }
-            unset($result['values']);
             // throw new \RuntimeException("Not implemented yet ! ".__CLASS__."::".__METHOD__);
             $result['ratio']= 'temporary';