+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 - Remove deprecated fallback property of Locale entity
 - Add a generateProductCompletenesses method on CompletenessGeneratorInterface, to generate completeness for one product
 - Add setCompletenesses and getCompletenesses method on ProductInterface and Product class
+- Add methods applyFilterByField, applySorterByField and update methods applyFilterByAttribute, applySorterByAttribute in ProductRepositoryInterface
+- Move findAllByAttributes and findOneByWithValues from FlexibleEntityRepositoryInterface to ProductRepositoryInterface
+- Move setFlexibleQueryBuilder, findAllByAttributes, findOneByWithValues, getFlexibleQueryBuilder, addJoinToValueTables, findAllByAttributesQB from FlexibleEntityRepository to ProductRepository (ORM)
 
 # 1.0.0 - "Hare We Go" (2014-03-06)